테스트용 액티비티 연결
